Reader: separate out card components
