+++ assets/stylesheets/_components.scss
@@ -100,6 +100,8 @@
 @import 'components/push-notification/style';
 @import 'components/rating/style';
 @import 'components/reader-full-post/style';
+@import 'components/reader-related-card/style';
+@import 'components/reader-search-card/style';
 @import 'components/related-posts/style';
 @import 'components/ribbon/style';
 @import 'components/search/style';
