+++ client/components/post-card/style.scss
@@ -1,403 +1,1 @@
-.post-card__site-info-title {
-	color: $gray;
-	float: left;
-	width: 100%;
-}
 
-.has-image .post-card__site-info-title {
-	width: calc( 100% - 120px );
-}
-
-.post-card__site-info {
-	color: $gray;
-	display: block;
-	font-size: 13px;
-	margin-right: 120px;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	width: 100%;
-
-	a {
-		color: $gray;
-		text-decoration: none;
-
-		&:visited {
-			color: inherit;
-		}
-
-		&:hover {
-			color: $blue-wordpress;
-		}
-	}
-
-	.site-icon {
-		align-self: auto;
-		float: left;
-		display: inline-block;
-		margin-right: 4px;
-	}
-}
-
-.post-card__title {
-	float: left;
-	line-height: 1.2;
-	margin-top: 6px;
-	width: 98%;
-}
-
-.post-card__anchor {
-	color: $gray-dark;
-	font-size: 15px;
-	font-weight: 600;
-	line-height: 1.4;
-	@extend %content-font;
-
-	&:visited {
-		color: $gray-dark;
-	}
-
-	&:hover {
-		color: $blue-wordpress;
-	}
-}
-
-.post-card__thumbnail {
-	border-radius: 2px;
-	clear: right;
-	float: right;
-	max-height: 68px;
-	max-width: 96px;
-}
-
-// Search cards
-.post-card__search.card {
-	position: relative;
-	padding: 20px;
-
-	&.has-thumbnail {
-		padding-left: 250px;
-
-		@include breakpoint( "660px-960px" ) {
-			padding-left: 135px;
-		}
-
-		@include breakpoint( "<480px" ) {
-			padding-left: 108px;
-		}
-	}
-
-	&.has-primary-follow-button.is-photo {
-
-		&::before{
-			opacity: .65;
-		}
-	}
-
-	// Search in CS
-	&.has-primary-follow-button .follow-button,
-	&.has-primary-follow-button.is-photo .follow-button {
-		background: none;
-		border-color: $blue-wordpress;
-		float: right;
-		position: relative;
-		z-index: 1;
-
-		.gridicon {
-			fill: $blue-wordpress;
-		}
-
-		.follow-button__label {
-			color: $blue-wordpress;
-		}
-
-		&:hover {
-			border-color: $gray-dark;
-
-			.gridicon {
-				fill: $gray-dark;
-			}
-
-			.follow-button__label {
-				color: $gray-dark
-			}
-		}
-
-		&.is-following {
-			border-color: $alert-green;
-
-			.gridicon {
-				fill: $alert-green;
-			}
-
-			.follow-button__label {
-				color: $alert-green;
-			}
-		}
-	}
-
-	&.has-primary-follow-button.is-photo .follow-button {
-		border-color: $white;
-
-		.gridicon {
-			fill: $white;
-		}
-
-		.follow-button__label {
-			color: $white;
-		}
-
-		&:hover {
-			border-color: lighten( $gray, 20% );
-
-			.gridicon {
-				fill: lighten( $gray, 20% );
-			}
-
-			.follow-button__label {
-				color: lighten( $gray, 20% );
-			}
-		}
-	}
-
-	&.is-photo {
-		min-height: 160px;
-		padding: 20px 20px 12px;
-
-		&::before{
-			background: $gray-dark;
-			content: '';
-			height: 100%;
-			opacity: .4;
-			position: absolute;
-				left: 0;
-				top: 0;
-			width: 100%;
-			z-index: 1;
-		}
-
-		.post-card__search-featured-image {
-			width: 100%;
-		}
-
-		.post-card__search-title {
-			display: block;
-			padding: 30px 20px 0 0;
-			position: relative;
-			z-index: 1;
-		}
-
-		.post-card__byline-author-and-site {
-			color: $white;
-		}
-
-		.post-card__search-title-link {
-			color: $white;
-
-			&:hover {
-				opacity: .8;
-			}
-		}
-
-		.post-card__search-byline {
-			position: relative;
-
-			.post-card__byline-link,
-			.post-card__byline-link:visited {
-				color: $white;
-			}
-
-			.gravatar {
-				border: 1px solid rgba( 255, 255, 255, 0.4 );
-			}
-		}
-
-		.post-card__author-and-site {
-			color: $white;
-		}
-
-		.post-card__search-social {
-			position: relative;
-		}
-
-		.post-card__search-excerpt {
-			display: none;
-		}
-
-		.gridicon {
-			fill: $white;
-		}
-
-		.comment-button__label-count,
-		.like-button__label-count {
-			color: $white;
-		}
-
-		.post-card__search-title-link,
-		.post-card__search-byline,
-		.post-card__search-social {
-			text-shadow: 0 1px rgba( 0, 0, 0, 0.3 );
-			z-index: 2;
-		}
-
-		.follow-button.is-following {
-
-			.gridicon {
-				fill: $white;
-			}
-
-			.follow-button__label {
-				color: $white;
-			}
-		}
-	}
-}
-
-.post-card__search-title {
-	display: inline;
-	@extend %content-font;
-	font-size: 20px;
-	line-height: 1.2;
-	font-weight: 700;
-
-	@include breakpoint( ">480px" ) {
-		font-size: 24px;
-		line-height: 28px;
-	}
-}
-
-.post-card__search .post-card__search-title-link {
-	color: inherit;
-
-	&:hover {
-		color: darken( $gray, 20% );
-	}
-
-	&:visited {
-		color: $gray;
-	}
-}
-
-.post-card__search-featured-image {
-	display: block;
-	position: absolute;
-		bottom: 0;
-		left: 0;
-		top: 0;
-	width: 230px;
-
-	@include breakpoint( "660px-960px" ) {
-		width: 115px;
-	}
-
-	@include breakpoint( "<480px" ) {
-		width: 90px;
-	}
-}
-
-.post-card__search-social {
-	display: flex;
-	float: right;
-	margin-bottom: 6px;
-	padding-left: 20px;
-}
-
-.post-card__search-byline {
-	font-size: 13px;
-	list-style: none;
-	margin: 6px 0;
-
-	.gravatar {
-		vertical-align: text-bottom;
-		margin: 0 4px 0 0;
-	}
-}
-
-.post-card__byline-author-and-site {
-	color: lighten( $gray, 10% );
-	margin-right: 6px;
-}
-
-.post-card__author-and-site {
-	color: lighten( $gray, 10% );
-	margin-right: 10px;
-}
-
-.post-card__search-excerpt {
-	@extend %content-font;
-	font-size: 14px;
-	margin-top: 6px;
-}
-
-// Custom styles for Follow button
-.post-card__search-byline .follow-button {
-	background: none;
-	border-style: none;
-	font-size: 13px;
-	margin: 0;
-	padding: 0;
-	position: relative;
-		top: -2px;
-
-	.gridicon {
-		fill: $blue-wordpress;
-	}
-
-	.follow-button__label {
-		color: $blue-wordpress;
-
-		@include breakpoint( "<660px" )
-			display: inline;
-	}
-
-	&:hover {
-
-		.gridicon {
-			fill: lighten( $gray, 10% );
-		}
-
-		.follow-button__label {
-			color: lighten( $gray, 10% );
-		}
-	}
-
-	&:focus {
-		border-radius: 0;
-	}
-
-	&.is-following {
-
-		.gridicon {
-			fill: $alert-green;
-		}
-
-		.follow-button__label {
-			color: $alert-green;
-		}
-
-		&:hover {
-
-			.gridicon {
-				fill: $blue-wordpress;
-			}
-
-			.follow-button__label {
-				color: $blue-wordpress;
-			}
-		}
-	}
-}
-
-// Custom colors for Site and Author info
-
-.reader__content {
-
-	.post-card__byline-link,
-	.post-card__byline-link:visited {
-		color: lighten( $gray, 10% );
-
-		&:hover {
-			color: $blue-wordpress;
-		}
-	}
-}
