+++ client/components/reader-search-card/index.jsx
@@ -9,7 +9,7 @@ import classnames from 'classnames';
  * Internal Dependencies
  */
 import Card from 'components/card';
-import AuthorAndSite from './author-and-site';
+import AuthorAndSite from 'components/reader-author-and-site';
 import FollowButton from 'reader/follow-button';
 import LikeButton from 'reader/like-button';
 import CommentButton from 'blocks/comment-button';
@@ -17,7 +17,7 @@ import DisplayTypes from 'state/reader/posts/display-types';
 
 function FeaturedImage( { image, href } ) {
 	return (
-		<a className="post-card__search-featured-image" href={ href } style={ {
+		<a className="reader-search-card__featured-image" href={ href } style={ {
 			backgroundImage: 'url(' + image.uri + ')',
 			backgroundSize: 'cover',
 			backgroundRepeat: 'no-repeat',
@@ -27,8 +27,8 @@ function FeaturedImage( { image, href } ) {
 
 function SearchByline( { post, site, feed, showFollowButton = true } ) {
 	return (
-		<div className="post-card__search-byline ignore-click">
-			<span className="post-card__search-byline-item">
+		<div className="reader-search-card__byline ignore-click">
+			<span className="reader-search-card__byline-item">
 				<AuthorAndSite post={ post } site={ site } feed={ feed } showGravatar={ true } />
 				{ showFollowButton && <FollowButton siteUrl={ post.site_URL } railcar={ post.railcar } /> }
 			</span>
@@ -43,7 +43,7 @@ export function SearchPostCard( { post, site, feed, onClick = noop, onCommentCli
 		length: isPhotoOnly ? 50 : 140,
 		separator: /,? +/
 	} );
-	const classes = classnames( 'post-card__search', {
+	const classes = classnames( 'reader-search-card', {
 		'has-thumbnail': !! featuredImage,
 		'is-photo': isPhotoOnly,
 		'has-primary-follow-button': showPrimaryFollowButton
@@ -52,21 +52,21 @@ export function SearchPostCard( { post, site, feed, onClick = noop, onCommentCli
 	return (
 		<Card className={ classes } onClick={ partial( onClick, { post, site, feed } ) }>
 		{ featuredImage && <FeaturedImage image={ featuredImage } href={ post.URL } /> }
-			{ ! showPrimaryFollowButton && <div className="post-card__search-social ignore-click">
+			{ ! showPrimaryFollowButton && <div className="reader-search-card__social ignore-click">
 				<CommentButton
 					commentCount={ post.discussion.comment_count }
 					tagName="span" showLabel={ false }
 					onClick={ onCommentClick }/>
 				<LikeButton siteId={ post.site_ID } postId={ post.ID } tagName="span" showZeroCount={ false } showLabel={ false } />
 			</div> }
-			{ showPrimaryFollowButton && <div className="post-card__search-follow-primary ignore-click">
+			{ showPrimaryFollowButton && <div className="reader-search-card__follow-primary ignore-click">
 				<FollowButton siteUrl={ post.site_URL } railcar={ post.railcar } />
 			</div> }
-			<h1 className="post-card__search-title">
-				<a className="post-card__search-title-link" href={ post.URL }>{ title }</a>
+			<h1 className="reader-search-card__title">
+				<a className="reader-search-card__title-link" href={ post.URL }>{ title }</a>
 			</h1>
 			<SearchByline post={ post } site={ site } feed={ feed } showFollowButton={ ! showPrimaryFollowButton } />
-			<div className="post-card__search-excerpt">{ post.short_excerpt }</div>
+			<div className="reader-search-card__excerpt">{ post.short_excerpt }</div>
 		</Card>
 	);
 }
